+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Add support for append user property operation.
+
 ## 2.3.0 (November 30, 2015)
 
 * Log if Google Play Services is enabled for the application.